backend/requirements.txt
alembic==1.14.1
asyncpg==0.30.0
boto3==1.37.11
fastapi==0.115.11
pydantic-settings==2.8.1
psycopg[binary]==3.2.6
sqlalchemy==2.0.38
uvicorn==0.34.0
websockets==15.0.1
